Code mostly cleaned up
