+++ x/tokenomics/keeper/keeper.go
@@ -30,6 +30,7 @@ type Keeper struct {
 	proofKeeper       types.ProofKeeper
 	sharedKeeper      types.SharedKeeper
 	sessionKeeper     types.SessionKeeper
+	serviceKeeper     types.ServiceKeeper
 
 	sharedQuerier client.SharedQueryClient
 }
@@ -47,6 +48,7 @@ func NewKeeper(
 	proofKeeper types.ProofKeeper,
 	sharedKeeper types.SharedKeeper,
 	sessionKeeper types.SessionKeeper,
+	serviceKeeper types.ServiceKeeper,
 ) Keeper {
 	if _, err := sdk.AccAddressFromBech32(authority); err != nil {
 		panic(fmt.Sprintf("invalid authority address: %s", authority))
@@ -67,6 +69,7 @@ func NewKeeper(
 		proofKeeper:       proofKeeper,
 		sharedKeeper:      sharedKeeper,
 		sessionKeeper:     sessionKeeper,
+		serviceKeeper:     serviceKeeper,
 
 		sharedQuerier: sharedQuerier,
 	}
